Use saltine/libsodium by default, and an opt-in flag to use shecretbox/cryptonite
Since the security of cryptonite nor shecretbox aren't guaranteed, we
default to depending on libsodium which has been audited numerous times
by many institutions. We add a conditional compilation flag and some
CPP macros so that users can choose to switch to the other implementation
if they accept the risks associated, and cannot link to libsodium for any
reason.