Remove unnecessary Maybe wrap/unwrap in oggPageExtractC
Previously we always wrapped `unconsumedBytes` with the Maybe
constructor, before performing prepending to `mbChunk`. This was
unnecessary computation if `unconsumedBytes` was `BS.empty` and
`mbChunk` was Nothing. The change adds an early if-conditional
for that. It also removes the Maybe wrapping, and instead uses
`fromMaybe` to unwrap the chunk from the conduit, which reduces
computation.