Use lazy bytestring within oggPageExtractC inner loop
In the inner loop for oggPageExtractC, we prepend the unconsumed bytes
with the new chunk. Using strict bytestrings incurs a O(m+n) memcpy
cost when performing the appending, whereas if we switch to lazy
bytestrings, we get two benefits:

1. We avoid BL.toStrict when recursing, since the `binary` package
   unconditionally uses Lazy bytestrings anyway.
2. We get a faster append, since lazy bytestring append is O(m) in
   the length of the spine in the first bytestring.

There are no places within the code that are affected negatively by
this change.

Ref:
https://hackage.haskell.org/package/bytestring-0.11.3.1/docs/src/Data.ByteString.Lazy.Internal.html#append
https://hackage.haskell.org/package/bytestring-0.11.3.1/docs/src/Data.ByteString.Internal.html#append