Relax upper bound on dependencies to support GHC 9.10
While discord-haskell is still awaiting for GHC 9.10 support to be released
on Hackage, we can already prepare other dependency bounds so that once
they've released it, it'll only involve one bound change.

With this commit, the project can now build on GHC 9.10 if you use the
latest git HEAD for discord-haskell instead of the Hackage version.