Implement ProbeCodec for `play`, which tries to avoid transcoding where possible
Using ProbeCodec as an argument for `play` makes discord-haskell-voice use the
provided ffprobe binary to find out whether the stream is already OPUS-encoded or
is raw PCM bytes. In either case, we can skip transcoding via ffmpeg and can use
Haskell functions to efficiently make Discord-friendly audio packets.

This is beneficial when users can afford to wait for a few seconds to avoid
transcoding.